# WIP: faster approximate spherical distance