@taldcroft 2023-10-04
Strangely, this seems to actually make `get_agasc_cone` run *slower* for AGASC 1p8 and make no different for AGASC 1p7. This makes no sense.
```
(ska3) ➜  dev git:(sphere-dist-fast) ✗ python profile_get_functions.py ../proseco_agasc_1p8rc4.h5  # original
AGASC module version: 4.16.1.dev36+gfdc28bf.d20231004
AGASC module file: /Users/aldcroft/git/agasc/agasc/__init__.py
AGASC file: ../proseco_agasc_1p8rc4.h5
agasc_get_cone: time for 500 queries: 3.95 s
get_star: time for 200 queries: 1.38 s
get_stars: time for 10 queries of 200 stars:  0.94 s

(ska3) ➜  dev git:(sphere-dist-fast) ✗ python profile_get_functions.py ../proseco_agasc_1p8rc4.h5  # fast
AGASC module version: 4.16.1.dev36+gfdc28bf.d20231004
AGASC module file: /Users/aldcroft/git/agasc/agasc/__init__.py
AGASC file: ../proseco_agasc_1p8rc4.h5
agasc_get_cone: time for 500 queries: 4.00 s
get_star: time for 200 queries: 1.41 s
get_stars: time for 10 queries of 200 stars:  0.91 s

(ska3) ➜  dev git:(sphere-dist-fast) ✗ python profile_get_functions.py ../proseco_agasc_1p7.h5  # original
AGASC module version: 4.16.1.dev36+gfdc28bf.d20231004
AGASC module file: /Users/aldcroft/git/agasc/agasc/__init__.py
AGASC file: ../proseco_agasc_1p7.h5
agasc_get_cone: time for 500 queries: 5.11 s
get_star: time for 200 queries: 1.37 s
get_stars: time for 10 queries of 200 stars:  0.89 s

(ska3) ➜  dev git:(sphere-dist-fast) ✗ python profile_get_functions.py ../proseco_agasc_1p7.h5  # fast
AGASC module version: 4.16.1.dev36+gfdc28bf.d20231004
AGASC module file: /Users/aldcroft/git/agasc/agasc/__init__.py
AGASC file: ../proseco_agasc_1p7.h5
agasc_get_cone: time for 500 queries: 6.57 s
get_star: time for 200 queries: 1.42 s
get_stars: time for 10 queries of 200 stars:  0.87 s
```