@taldcroft 2023-10-14
Closing this. If we ever need a fast sphere_dist then maybe we'll remember, but for `get_agasc_cone` the sphere_dist part is a negligible